+++ sources/Renderer/TextureUtils.cpp
@@ -93,7 +93,7 @@ LLGL_EXPORT SubresourceFootprint CalcPackedSubresourceFootprint(
         const auto mipExtent = GetMipExtent(type, extent, mipLevel);
         const auto numLayers = mipExtent.depth * numArrayLayers;
         footprint.rowAlignment  = alignment;
-        footprint.rowSize       = GetMemoryFootprint(format, mipExtent.width);
+        footprint.rowSize       = static_cast<std::uint32_t>(GetMemoryFootprint(format, mipExtent.width));
         footprint.rowStride     = GetAlignedSize(footprint.rowSize, alignment);
         footprint.layerSize     = (mipExtent.height > 1 ? footprint.rowStride * (mipExtent.height - 1) + footprint.rowSize : footprint.rowSize * mipExtent.height);
         footprint.layerStride   = footprint.rowStride * mipExtent.height;
